chore(compiler-ts): consolidate type definitions into src/types/index.ts (#180)
Move every hand-written type this package exposes (WorkflowType,
ActorReference, RegisteredActor, WrittenActor, FsmPluginValidationResult,
FsmDraftStateNode, etc.) out of the files that happened to define them
first (util.ts, operation-logic-scaffold.ts, generate-fsm-json.ts) into
one entry point, src/types/index.ts. All 14 consumer files (13 source +
the public barrel) now import types from there instead of scattered
across implementation files.

types/index.ts also re-exports the schema-derived types this package
actually uses (FsmMachineJson, ActionObject, AtomicStateNode,
CompoundStateNode, FinalStateNode, HistoryStateNode, ParallelStateNode)
from database-src/generated/fsm-machine-schema.types.ts — by name, not a
blanket `export type *`, so only types something in this package actually
imports get re-exported; add to the list when a new one is needed rather
than importing database-src directly. InvokeObject stays a direct
internal import (only used for FsmDraftInvoke's own definition, never
re-exported).

Verified the cross-package import doesn't break @pgfsm/compiler's dnt
npm-publish build — the type-only import gets bundled self-contained into
dist/ rather than erased, and the build succeeds either way.
