+++ packages/fsm-compiler-ts/src/validate-async-operation-logic.ts
@@ -7,17 +7,14 @@ const _checkerDir = new URL("./checkers", import.meta.url).pathname;
 let _rustCheckerBin: string | null | false = null;
 let _goCheckerBin: string | null | false = null;
 
-import {
-  operationFileExtension,
-  type OperationLang,
-} from "./operation-logic-scaffold.ts";
-import {
-  type ActorPluginValidationResult,
-  type ActorReference,
-  DenoCommand,
-  isVersionFolderName,
-  type WorkflowType,
-} from "./util.ts";
+import { operationFileExtension } from "./operation-logic-scaffold.ts";
+import { DenoCommand, isVersionFolderName } from "./util.ts";
+import type {
+  ActorPluginValidationResult,
+  ActorReference,
+  OperationLang,
+  WorkflowType,
+} from "./types/index.ts";
 
 /**
  * Every actor-language checker below shells out to that language's runtime
