+++ packages/fsm-compiler-ts/test/validate-fsm-plugin-load.test.ts
@@ -4,7 +4,7 @@ import {
   isFunction,
   validateSyncOperationFromFolder,
 } from "../src/validate-sync-operation-logic.ts";
-import type { FsmMachineJson } from "../../database-src/generated/fsm-machine-schema.types.ts";
+import type { FsmMachineJson } from "../src/types/index.ts";
 
 // isFunction
 Deno.test("isFunction - returns true for functions", () => {
